+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 
 import {
   camelizeCommand,
+  clearEmptyFolderCommand,
   createIndexCommand,
   createIndexOnlyTargetCommand,
   formatSortTsCommand,
@@ -17,18 +18,20 @@ import { extensionNamespace } from './utils/extensionNamespace';
 export function activate(context: vscode.ExtensionContext) {
   OutputChannel.appendLine(`"${extensionNamespace}" is now active!`, false);
 
-  context.subscriptions.push(openAndRunCommand);
-  context.subscriptions.push(formatSortTsCommand);
-  context.subscriptions.push(formatSortTsOnceCommand);
-  context.subscriptions.push(prettierFormatCommand);
-  context.subscriptions.push(toTsCommand);
-  context.subscriptions.push(camelizeCommand);
-  context.subscriptions.push(switchTypeInterfaceCommand);
-
-  context.subscriptions.push(createIndexCommand('js'));
-  context.subscriptions.push(createIndexOnlyTargetCommand('js'));
-  context.subscriptions.push(createIndexCommand('both'));
-  context.subscriptions.push(createIndexOnlyTargetCommand('both'));
+  context.subscriptions.push(
+    openAndRunCommand,
+    clearEmptyFolderCommand,
+    formatSortTsCommand,
+    formatSortTsOnceCommand,
+    prettierFormatCommand,
+    toTsCommand,
+    camelizeCommand,
+    switchTypeInterfaceCommand,
+    createIndexCommand('js'),
+    createIndexOnlyTargetCommand('js'),
+    createIndexCommand('both'),
+    createIndexOnlyTargetCommand('both'),
+  );
 }
 
 // this method is called when your extension is deactivated